requirements.txt
# Requerimientos para el proyecto

streamlit>=1.0.0
sqlmodel>=0.0.31
sqlalchemy>=2.0.0
pydantic>=2.0.0
python-dotenv>=1.0.0
psycopg2-binary>=2.9.0  # Driver para PostgreSQL
